numa_avatar/mvp/requirements.txt
openai>=1.0.0
httpx>=0.23.0
pyyaml>=6.0